refactor: 쿼리 키 중앙 집중식 관리 구조 도입(#35)
- React Query 쿼리 키 관리를 개선하기 위해 중앙 집중식 관리 패턴을 도입하여 쿼리 키의 일관성과 유지보수성을 향상
