+++ src/api/index.js
@@ -1,5 +1,6 @@
-export * from './fetchMovieDetails';
-export * from './fetchNowPlayingMovies';
-export * from './fetchPopularMovieList';
-export * from './fetchSearchMovies';
-export * from './supabase';
+export * from "./fetchMovieDetails";
+export * from "./fetchNowPlayingMovies";
+export * from "./fetchPopularMovieList";
+export * from "./fetchSearchMovies";
+export * from "./queryKeys";
+export * from "./supabase";
